pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proxy-manager"
version = "1.5.3"
description = "HAProxy configuration management service with web UI."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "jinja2>=3.1.0",
    "bcrypt>=5.0.0",
    "pydantic>=2.0.0",
    "PyJWT>=2.8.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.27",
    "ruff>=0.6",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["proxy_manager*"]

[tool.setuptools.package-data]
proxy_manager = ["static/**/*", "templates/**/*"]

[tool.ruff]
target-version = "py313"
line-length = 180

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "S", "SIM"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S105", "S106"]

[tool.ruff.lint.isort]
known-first-party = ["proxy_manager"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "def __repr__",
]

[tool.pyright]
pythonVersion = "3.13"
pythonPlatform = "Linux"
typeCheckingMode = "strict"
include = ["proxy_manager"]
exclude = ["tests", "alembic", "**/__pycache__"]
reportMissingTypeStubs = false
reportMissingImports = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnknownLambdaType = false
reportCallInDefaultInitializer = false
reportImplicitOverride = false
reportUntypedFunctionDecorator = false
reportUntypedBaseClass = false
reportInvalidTypeForm = false
